wrapfs: use el7 inode_operations_wrapper for compatibility with underlying fs
redhat7 uses an inode_operations_wrapper struct to backport
some inode operations from newer kernels. The presence of this
extended struct is indicated by 'inode->i_flags & S_IOPS_WRAPPER'.

When the S_IOPS_WRAPPER flag is not aligned with the layout
of the iops struct, the vfs_rename-call will read an invalid
function pointer and crash the kernel.

The crash does only happen when the virtual filesystem is mounted
on top of a filesystem using the S_IOPS_WRAPPER layout (btrfs,cifs,
ext2,ext4,fuse,gfs2,xfs).

This crash happens also with the original wrapfs implementation
and the ecryptfs module included in the redhat kernel tree. The
redhat7 overlayfs kernel module is compatible with the wrapper layout.

see: redhat-7.9 kernel sources
     https://vault.centos.org/7.9.2009/os/Source/SPackages/kernel-3.10.0-1160.el7.src.rpm